Fix for django bug, where : and , symbols are not properly handled
When saving package files, ':' and ',' symbols were omitted from the
filename, even when they were passed properly. This code substitutes them
for their hexadecimal value.

Also this code changes download URLs to a safer version (names are hashed
with sha256, so they are unguessable). This only happens when DEBUG is
set to False

Change-Id: Iba2b52c6aef0b416ac7a2c276aa0ae72904be70d
Fixes: AUTOSUITE-1450
Reviewed-by: Egor Nemtsev <enemtsev@luxoft.com>